#922: Moved redis classes to Abp.Runtime.Caching.Redis
