+++ src/Abp.RedisCache/Runtime/Caching/Redis/AbpRedisCacheModule.cs
@@ -1,7 +1,7 @@
-﻿using Abp.Modules;
-using System.Reflection;
+﻿using System.Reflection;
+using Abp.Modules;
 
-namespace Abp.RedisCache
+namespace Abp.Runtime.Caching.Redis
 {
     /// <summary>
     /// This modules is used to replace ABP's cache system with Redis server.
