+++ src/Abp.RedisCache/Runtime/Caching/Redis/RedisCacheConfigurationExtensions.cs
@@ -1,0 +1,7 @@
+﻿namespace Abp.Runtime.Caching.Redis
+{
+    public static class RedisCacheConfigurationExtensions
+    {
+
+    }
+}
